feat: also display withdrawals in user activity
